Added Ability To Create An Existing PG On Another vSwitch
Added the ability to create a port group with the same name on another
vSwitch. The current process is to detect if the port group already
exists, if it does create a new port group with -ToBeRemoved added to
it. After creating this new port group any existing vms that reside on
on the port group to be created are moved to the -ToBeRemoved port group
temporarily. Once all vms are moved the original port group will be
removed and then recreated on another vSwitch. Once the new port group
is added all vms are then moved from the -ToBeRemoved port group over to
the new port group. The -ToBeRemoved port group ends up getting removed
on another run of the role.

* To fix still: Add the ability to recreate the same port group on the
original vSwitch and reverse the process. Currently this is not working.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>