Fix VLAN Management For Port Groups
There was an issue with VLANs being properly defined when moving
port groups between vSwitches. The scenario was that when a port group
already existed and the port group needed to be placed onto a different
vSwitch. The port group would move correctly, however the VLAN id was
not being properly defined. As well as once a port group was moved there
was not an additional logic check on consecutive runs that would
validate that a port group VLAN was defined properly.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>