Refactored Bootstrap VM Deployments
Converted win_shell with inline code to a win_template and win_shell to
actually execute the script. This seemed to make most sense as the
logic was/is growing. Again, this was also done for the network
scripting as I hit a limit on the size of the inline code that I could
send via Ansible.

Added logic to determine if VM UUID exists in inventory. If it does and
the deploy script detects that a VM which should be deployed has been
renamed in vSphere then it will compate inventory UUID to the actual VM
UUID captured from vSphere. If the UUIDs match then the VM in vSphere
will be changed to match the inventory. If the inventory UUID does not
exist the script will assume that the VM is a new one and deploy it
again. Currently there is not any logic to detect if they are similar.

Added logic to determine if the VM in vSphere is on the correct network
as defined in the inventory. If it does not then the network for the VM
in vSphere will be changed. If the VM is powered on then the network
will be disconnected and then reconnected. This assists in the renewal
of IP address if using DHCP (Recommended).

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>