Replace PowerDNS VMs/Tasks As DDI
Replaced all VMs and tasks related to PowerDNS to now be DDI related.
This is a more clear distinguished methodology. The DDI VMs are
configured as static IP addresses and there is no longer a requirement
for DHCP to be available on the network to provision these VMs. The
requirement for VIPs to be configured for DNS is no longer required as
well due to the fact that the DDI VMs are statically assigned addressed
now.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>