Add VM Autostart Policies For DDI/LB VMs
Added logic in order to properly set VM Autostart policies. The current
method is to assume that the DDI VMs have priority. So they are defined
in order based on Ansible group along with Jinja2 loop.index to define
them in order. The LB VMs are incremented based on the DDI VMs unless
the DDI VMs do not exist and then they increment from 1.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>