Fixed Issue With VM Guest Family Detection
There was an issue when deploying VMs which did not detect the VM Guest
Family in time to properly configure the VM hostname. The VM Guest
Family would show as null therefore skipping the hostname change. This
is now a loop until the VM Guest Family is not null before proceeding.
Also cleaned up some formatting indentation.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>