Added check to determine if VMs have already been deployed.
When kicking off a task such as vsphere_vms_info when a set of VMs were
not deployed would result in an error because it was looking for the VMs
and assuming that they already exist. Now it adds VMs not already
deployed to a group not_deployed and uses this group to do additional
logic for the remaining tasks.